SEQUOIADBMAINSTREAM-6479 去除编译时的 warning

+++ SequoiaDB/engine/client/clientcpp.cpp
@@ -8857,7 +8857,7 @@ do                                                            \
       _sock->quickAck() ;
 
       ossEndianConvertIf4 ( length, realLen, _endianConvert ) ;
-      if ( realLen < sizeof(MsgOpReply) )
+      if ( (UINT32)realLen < sizeof(MsgOpReply) )
       {
          rc = SDB_NET_BROKEN_MSG ;
          goto error ;
